requirements.txt
wheel~=0.43.0
celery~=5.4.0
redis~=5.0.7
setuptools~=68.2.0
PyJWT~=2.8.0
fastapi~=0.111.0
SQLAlchemy~=2.0.31
bcrypt~=4.1.3
pydantic~=2.7.4
vertexai~=1.49.0
psycopg2-binary~=2.9.9
alembic~=1.13.2
uvicorn~=0.30.1